requirements.txt
torch==2.2.2+cpu
torchvision==0.17.2+cpu
numpy==1.26.4
pillow==10.3.0
cairosvg==2.7.1
tqdm==4.66.2
